提交

+++ workspace/manager/src/main/java/com/xiaoyi/manager/service/impl/TeachingResourceServiceImpl.java
@@ -322,7 +322,7 @@ public class TeachingResourceServiceImpl implements ITeachingResourceService {
 				
 				//record.setTeacherid(teacherId);
 				
-				int operateNum = params.getIntValue("operateNum");
+				float operateNum = params.getFloatValue("operateNum");
 				int operateType = params.getIntValue("operateType");
 				
 				//课时操作
